api/requirements.txt
fastapi
uvicorn
pyjwt
requests
